Changed API structure
