+++ nginx/nginx.conf
@@ -3,6 +3,12 @@ events {}
 http {
     include /etc/nginx/mime.types;
 
+    map $http_origin $cors_origin {
+        default "";
+        "~^https?://localhost(:[0-9]+)?$" $http_origin;
+        "~^https?://127\.0\.0\.1(:[0-9]+)?$" $http_origin;
+    }
+
     server {
         listen 80;
         server_name _;
@@ -15,6 +21,22 @@ http {
         }
 
         location /api/ {
+            if ($request_method = OPTIONS) {
+                add_header Access-Control-Allow-Origin $cors_origin always;
+                add_header Vary Origin always;
+                add_header Access-Control-Allow-Methods "GET, POST, PUT, DELETE, OPTIONS" always;
+                add_header Access-Control-Allow-Headers "Content-Type, Authorization" always;
+                add_header Access-Control-Max-Age 86400 always;
+                add_header Content-Length 0;
+                add_header Content-Type "text/plain; charset=UTF-8";
+                return 204;
+            }
+
+            add_header Access-Control-Allow-Origin $cors_origin always;
+            add_header Vary Origin always;
+            add_header Access-Control-Allow-Methods "GET, POST, PUT, DELETE, OPTIONS" always;
+            add_header Access-Control-Allow-Headers "Content-Type, Authorization" always;
+
             proxy_pass http://api:8000/;
             proxy_http_version 1.1;
             proxy_set_header Host $host;
