+++ web-v2/.env.example
@@ -1,3 +1,3 @@
-VITE_API_URL=http://localhost:8000
-VITE_ESPLORA_BASE_URL=https://blockstream.info/liquidtestnet/api
+VITE_API_URL=http://localhost:80
+VITE_ESPLORA_BASE_URL=https://blockstream.info/liquidtestnet
 VITE_NETWORK=liquidtestnet
