+++ web-v2/package.json
@@ -19,6 +19,7 @@
     "lwk_web": "file:../lwk_wasm/pkg_web",
     "@heroui/react": "^3.0.4",
     "@tanstack/react-query": "^5.100.10",
+    "axios": "^1.7.0",
     "react": "^19.2.6",
     "react-dom": "^19.2.6",
     "react-router-dom": "^7.15.0",
