--- web-v2/src/api/hooks/index.ts
@@ -1,3 +1,0 @@
-export * from './esplora'
-export * from './offers'
-export { GC_TIME_MS, STALE_TIME_MS } from './staleTime'
