--- web-v2/src/api/index.ts
@@ -1,8 +1,0 @@
-export * from './errors'
-export * from './esplora'
-export * from './hooks'
-export * from './indexer'
-export * from './queryKeys'
-export * from './retry'
-export * from './schemas'
-export * from './transport'
