+++ web-v2/src/api/indexer/hooks.ts
@@ -1,5 +1,6 @@
 import { useQuery, type UseQueryResult } from '@tanstack/react-query'
 
+import { GC_TIME_MS, STALE_TIME_MS } from '../staleTime'
 import {
   fetchOffer,
   fetchOfferIdsByBorrowerPubkey,
@@ -9,14 +10,13 @@ import {
   fetchOffers,
   fetchOfferUtxos,
   type ListOffersParams,
-} from '../indexer'
-import { queryKeys } from '../queryKeys'
-import type { OfferDetails, OfferParticipant, OfferShort, OfferUtxo } from '../schemas'
-import { GC_TIME_MS, STALE_TIME_MS } from './staleTime'
+} from './methods'
+import { offersQueryKeys } from './queryKeys'
+import type { OfferDetails, OfferParticipant, OfferShort, OfferUtxo } from './schemas'
 
 export function useOffers(params: ListOffersParams = {}): UseQueryResult<OfferShort[]> {
   return useQuery({
-    queryKey: queryKeys.offers.list(params),
+    queryKey: offersQueryKeys.list(params),
     queryFn: ({ signal }) => fetchOffers(params, { signal }),
     staleTime: STALE_TIME_MS.medium,
   })
@@ -24,7 +24,7 @@ export function useOffers(params: ListOffersParams = {}): UseQueryResult<OfferSh
 
 export function useOffer(offerId: string): UseQueryResult<OfferDetails> {
   return useQuery({
-    queryKey: queryKeys.offers.detail(offerId),
+    queryKey: offersQueryKeys.detail(offerId),
     queryFn: ({ signal }) => fetchOffer(offerId, { signal }),
     staleTime: STALE_TIME_MS.realtime,
     enabled: !!offerId,
@@ -33,7 +33,7 @@ export function useOffer(offerId: string): UseQueryResult<OfferDetails> {
 
 export function useOfferUtxos(offerId: string): UseQueryResult<OfferUtxo[]> {
   return useQuery({
-    queryKey: queryKeys.offers.utxos(offerId),
+    queryKey: offersQueryKeys.utxos(offerId),
     queryFn: ({ signal }) => fetchOfferUtxos(offerId, { signal }),
     staleTime: STALE_TIME_MS.realtime,
     enabled: !!offerId,
@@ -42,7 +42,7 @@ export function useOfferUtxos(offerId: string): UseQueryResult<OfferUtxo[]> {
 
 export function useOfferParticipants(offerId: string): UseQueryResult<OfferParticipant[]> {
   return useQuery({
-    queryKey: queryKeys.offers.participants(offerId),
+    queryKey: offersQueryKeys.participants(offerId),
     queryFn: ({ signal }) => fetchOfferParticipants(offerId, { signal }),
     staleTime: STALE_TIME_MS.realtime,
     enabled: !!offerId,
@@ -51,7 +51,7 @@ export function useOfferParticipants(offerId: string): UseQueryResult<OfferParti
 
 export function useOfferParticipantsHistory(offerId: string): UseQueryResult<OfferParticipant[]> {
   return useQuery({
-    queryKey: queryKeys.offers.participantsHistory(offerId),
+    queryKey: offersQueryKeys.participantsHistory(offerId),
     queryFn: ({ signal }) => fetchOfferParticipantsHistory(offerId, { signal }),
     staleTime: STALE_TIME_MS.realtime,
     gcTime: GC_TIME_MS.long,
@@ -61,7 +61,7 @@ export function useOfferParticipantsHistory(offerId: string): UseQueryResult<Off
 
 export function useOfferIdsByScript(scriptPubkeyHex: string): UseQueryResult<string[]> {
   return useQuery({
-    queryKey: queryKeys.offers.byScript(scriptPubkeyHex),
+    queryKey: offersQueryKeys.byScript(scriptPubkeyHex),
     queryFn: ({ signal }) => fetchOfferIdsByScript(scriptPubkeyHex, { signal }),
     staleTime: STALE_TIME_MS.realtime,
     enabled: !!scriptPubkeyHex,
@@ -70,7 +70,7 @@ export function useOfferIdsByScript(scriptPubkeyHex: string): UseQueryResult<str
 
 export function useOfferIdsByBorrowerPubkey(borrowerPubkeyHex: string): UseQueryResult<string[]> {
   return useQuery({
-    queryKey: queryKeys.offers.byBorrower(borrowerPubkeyHex),
+    queryKey: offersQueryKeys.byBorrower(borrowerPubkeyHex),
     queryFn: ({ signal }) => fetchOfferIdsByBorrowerPubkey(borrowerPubkeyHex, { signal }),
     staleTime: STALE_TIME_MS.realtime,
     enabled: !!borrowerPubkeyHex,
